Removing user submitted password
Service is never run on password protected files because of filetypes.
Encrypted excel appears as document/office/passwordprotected and so
is never analyzed by the service. Extract will decrypt password protected
files and allow the decrypted version to be analyzed.
User passwords can be submitted in the Extract password field.